Massive reorganization. Pushing this up for incremental addition to the main project <3
