+++ src/components/layouts/CardLayout.tsx
@@ -1,17 +1,30 @@
 import { Content, ContentType } from "../../types/documents";
 import React from "react";
 import { Card } from "react-bootstrap";
-import { Atom, Molecule } from "./Layouts";
 import { DynamicMolecule } from "./molecules";
+import { Atom, Molecule } from "../../types/layouts";
 
+// The props for the base Card Layout
 interface CardLayoutProps {
   content: Content;
   contentType: ContentType;
   rules: any;
 }
 
+// The props for the Card Layout Header and Body.
+interface CardLayoutChildProps {
+  contentType: ContentType;
+  data: {
+    content: Content;
+  }
+}
 
-function CardLayoutHeader(props: any) {
+/**
+ * Renders the header of a content card
+ * @param props.contentType The content type that contains the layout
+ * @param props.data Contains all dynamic data that may be used in the header
+ */
+function CardLayoutHeader(props: CardLayoutChildProps) {
   const headerItems: JSX.Element[] = [];
   props.contentType.layout.header.forEach((molecule: Molecule | Atom) => {
     headerItems.push(
@@ -28,7 +41,12 @@ function CardLayoutHeader(props: any) {
   )
 }
 
-function CardLayoutBody(props: any) {
+/**
+ * Renders the header of a content card
+ * @param props.contentType The content type that contains the layout
+ * @param props.data Contains all dynamic data that may be used in the header
+ */
+function CardLayoutBody(props: CardLayoutChildProps) {
   const bodyItems: JSX.Element[] = [];
   props.contentType.layout.body.forEach((molecule: Molecule | Atom) => {
     bodyItems.push(
@@ -46,11 +64,18 @@ function CardLayoutBody(props: any) {
   );
 }
 
+/**
+ * Renders the full card layout
+ *
+ * @param props.content The specific content that we're rendering
+ * @param props.contentType The type of content that we're rendering. Contains the layout
+ * @param props.rules Any additional data that is global for the gamesystem
+ */
 export function CardLayout(props: CardLayoutProps) {
   const data = {
     content: props.content
   };
-  // console.log(props)
+  
   return (
     <Card>
       <CardLayoutHeader {...props} data={data}/>
