+++ src/components/layouts/atoms/AtomError.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Col } from "react-bootstrap";
+
+export function AtomError(props: any) {
+  return (
+    <Col {...(props.w || {}) } style={{color: "red", fontWeight: "bold"}}>
+      {props.errors}
+    </Col>
+  )
+}
