+++ src/components/layouts/atoms/Multiselect.tsx
@@ -1,19 +1,27 @@
 import React from "react";
 import { Col, Form } from "react-bootstrap";
-import { renderDisplayClasses } from ".";
+import { findValue, renderDisplayClasses } from "../../../utilities/layouts/atoms";
+import { AtomProps } from "../../../types/layouts/atom";
 import { Multiselect } from "../../design/forms/Forms";
-import { RenderError } from "../Layouts";
+import { AtomError } from "../atoms/AtomError";
 
-export function MultiselectInput(props: any) {
+/**
+ * Renders an input that allows for selecting multiple options
+ * @param props.atom The definition of the atom to render
+ * @param props.data The dynamic data that may be referenced by the atom
+ */
+export function MultiselectInput(props: AtomProps) {
   let errors: string = "";
-  const classes = renderDisplayClasses(props.atom.display);
-  const inputName = props.atom.staticValues.inputName || undefined;
-  const label = props.atom.staticValues.label || undefined;
-  const placeholder = props.atom.staticValues.placeholder || undefined;
-  
-  if (!inputName) { errors += "An input name is required."; }
 
-  if (errors) { return <RenderError w={props.atom.w} errors={errors}/>; }
+  // Ensures that we have an input name, since formik breaks without it
+  const inputName = findValue("inputName", props); 
+  if (!inputName) { errors += "An input name is required."; }
+  if (errors) { return <AtomError w={props.atom.w} errors={errors}/>; }
+
+  // Finds these values after to prevent unneeded usage of 
+  const classes = renderDisplayClasses(props.atom.display);
+  const label = findValue("label", props);
+  const placeholder = findValue("placeholder", props);
 
   return (
     <Form.Group as={Col} className={classes} {...(props.atom.w || {})}>
