+++ src/components/layouts/atoms/TextInput.tsx
@@ -1,26 +1,27 @@
 import React from "react";
 import { Col, Form } from "react-bootstrap";
-import { findValue } from ".";
+import { AtomProps } from "../../../types/layouts/atom";
+import { findValue } from "../../../utilities/layouts/atoms";
 import { Error, Input } from "../../design/forms/Forms";
-import { Atom, RenderError } from "../Layouts";
-
-interface TextInputProps {
-  atom: Atom;
-}
+import { AtomError } from "./AtomError";
 
 /**
- * Renders a text input
- * @param props 
+ * Renders a text input that connects with a Formik form
+ *
+ * @param props.atom The definition of the atom to render
+ * @param props.data The dynamic data that may be referenced by the atom
  */
-export function TextInput(props: TextInputProps) {
+export function TextInput(props: AtomProps) {
   let errors: string = "";
+  
+  // Ensures that inputName value is present as this will break 
+  // TODO - can we have this as a shared function?
   const inputName = findValue("inputName", props);
+  if (!inputName) { errors += "An input name is required."; }
+  if (errors) { return <AtomError w={props.atom.w} errors={errors}/>; }
+
   const label = findValue("label", props);
   const placeholder = findValue("placeholder", props);
-  
-  if (!inputName) { errors += "An input name is required."; }
-
-  if (errors) { return <RenderError w={props.atom.w} errors={errors}/>; }
 
   return (
     <Form.Group as={Col} {...(props.atom.w || {})}>
@@ -30,7 +31,6 @@ export function TextInput(props: TextInputProps) {
         placeholder={placeholder}
       />
       <Error name={inputName} />
-      
     </Form.Group>
   );
 }
