+++ src/types/enums/moleculeType.ts
@@ -1,0 +1,7 @@
+// Describes the different kinds of molecule types that can be rendered around a 
+// collection of atoms
+export enum MoleculeType {
+  Div = 0,
+  
+  Accordion,
+}
